POST records

+++ backend/src/main/java/com/manadev/dspesquisa/repositories/GameRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.manadev.dspesquisa.entities.Game;
 
 @Repository
-public interface GameRepository extends JpaRepository<Game, Integer> {
+public interface GameRepository extends JpaRepository<Game, Long> {
 
 }
